Olap kp dev (#72)
* add parallel query code and mtr test code (#51)

* MySQL-AP并行查询需求开发
1、支持TPCH Q4
2、并行查询开关修改

* 并行查询开关代码检视意见修改
